requirements.txt
google-cloud-aiplatform==1.74.0
anthropic[vertex]==0.40.0
litellm[proxy]==1.54.0
google-cloud-secret-manager>=2.16.4
arize-otel>=0.0.1
openinference-instrumentation-litellm>=0.0.1
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation-httpx==0.50b0
httpx<0.28
python-json-logger>=2.0.0
